requirements.txt
streamlit==1.37.1
pandas==2.2.2
PyPDF2==3.0.1
qdrant-client==1.10.1
sentence-transformers==3.0.1
langchain==0.2.10
langchain-community==0.2.9
langchain-google-genai==1.0.7
SQLAlchemy==2.0.32
python-dotenv==1.0.1